Simplify listener abstraction
